page

+++ back/compte.php
@@ -10,10 +10,16 @@ if (isset($_SESSION["id_compte"])) {
         //on switche sur la valeur contenue dans $_GET['action']
         switch ($_GET["cas"]) {
             case "inserer_compte":
+                if (isset($_POST['statut_compte'])){
+                    $select[$_POST['statut_compte']]="selected";
+                }
                 if (empty($_POST["nom_compte"])) {
                     $confirmation =
                         "<p class=\"pas_ok\">Le nom est obligatoire</p>";
-                } elseif (empty($_POST["email_compte"])) {
+                } elseif (empty($_POST["statut_compte"])) {
+                    $confirmation =
+                        "<p class=\"pas_ok\">Le statut est obligatoire</p>";}
+                elseif (empty($_POST["email_compte"])) {
                     $confirmation =
                         "<p class=\"pas_ok\">L'email est obligatoire</p>";
                 } elseif (empty($_POST["login_compte"])) {
@@ -23,23 +29,16 @@ if (isset($_SESSION["id_compte"])) {
                     $confirmation =
                         "<p class=\"pas_ok\">Le mot de passe est obligatoire</p>";
                 } else {
+
                     //on enregistre le compte dans la table compte
                     $requete =
                         "INSERT INTO comptes SET nom_compte='" .
-                        security($_POST["nom_compte"]) .
-                        "',
-                                                      prenom_compte='" .
-                        security($_POST["prenom_compte"]) .
-                        "',  
-                                                      email_compte='" .
-                        security($_POST["email_compte"]) .
-                        "',  
-                                                      login_compte='" .
-                        security($_POST["login_compte"]) .
-                        "',  
-                                                      pass_compte=SHA1('" .
-                        $_POST["pass_compte"] .
-                        "')";
+                        security($_POST["nom_compte"]) ."',
+                        prenom_compte='" .security($_POST["prenom_compte"]) ."',
+                        statut_compte='" .($_POST["statut_compte"]) ."',
+                        email_compte='" .security($_POST["email_compte"]) . "',  
+                        login_compte='" .security($_POST["login_compte"]) . "',  
+                        pass_compte=SHA1('" .$_POST["pass_compte"] ."')";
 
                     $resultat = mysqli_query($connexion, $requete);
 
@@ -57,16 +56,9 @@ if (isset($_SESSION["id_compte"])) {
                 break;
             case "avertir_compte":
                 if (isset($_GET["id_compte"])) {
-                    $confirmation =
-                        "<p>Voulez-vous supprimer le compte n°" .
-                        $_GET["id_compte"] .
-                        "</p>";
-                    $confirmation .=
-                        "<a href=\"back.php?action=compte&cas=supprimer_compte&id_compte=" .
-                        $_GET["id_compte"] .
-                        "\">OUI</a>&nbsp;&nbsp;&nbsp;";
-                    $confirmation .=
-                        "<a href=\"back.php?action=compte\">NON</a>";
+                    $confirmation = "<p>Voulez-vous supprimer le compte n°" . $_GET["id_compte"] . "</p>";
+                    $confirmation .= "<a href=\"back.php?action=compte&cas=supprimer_compte&id_compte=" . $_GET["id_compte"] . "\">OUI</a>&nbsp;&nbsp;&nbsp;";
+                    $confirmation .= "<a href=\"back.php?action=compte\">NON</a>";
                 }
 
                 break;
@@ -79,43 +71,36 @@ if (isset($_SESSION["id_compte"])) {
                     $ligne = mysqli_fetch_object($resultat);
                     //si c'est le dernier compte de la table (1 seule ligne dans la table)
                     if ($ligne->nb_compte == 1) {
-                        $confirmation =
-                            "<p class=\"pas_ok\">Le dernier compte autorisé ne peut pas être supprimé</p>";
+                        $confirmation = "<p class=\"pas_ok\">Le dernier compte autorisé ne peut pas être supprimé</p>";
                     } else {
-                        $requete2 =
-                            "DELETE FROM comptes WHERE id_compte='" .
-                            $_GET["id_compte"] .
-                            "'";
+                        $requete2 = "DELETE FROM comptes WHERE id_compte='" . $_GET["id_compte"] . "'";
                         $resultat2 = mysqli_query($connexion, $requete2);
-                        $confirmation =
-                            "<p class=\"ok\">Le compte a bien été supprimé</p>";
+                        $confirmation = "<p class=\"ok\">Le compte a bien été supprimé</p>";
                     }
                 }
 
                 break;
 
             case "recharger_compte":
-                $action_form =
-                    "modifier_compte&id_compte=" . $_GET["id_compte"];
+                $action_form = "modifier_compte&id_compte=" . $_GET["id_compte"];
                 if (isset($_GET["id_compte"])) {
-                    $requete =
-                        "SELECT * FROM comptes WHERE id_compte='" .
-                        $_GET["id_compte"] .
-                        "'";
+                    $requete = "SELECT * FROM comptes WHERE id_compte='" . $_GET["id_compte"] . "'";
                     $resultat = mysqli_query($connexion, $requete);
                     $ligne = mysqli_fetch_object($resultat);
                     $_POST["nom_compte"] = $ligne->nom_compte;
                     $_POST["prenom_compte"] = $ligne->prenom_compte;
                     $_POST["email_compte"] = $ligne->email_compte;
                     $_POST["login_compte"] = $ligne->login_compte;
+                    if (isset($ligne->statut_compte)){
+                        $select[$ligne->statut_compte]="selected";
+                    }
                 }
                 break;
 
             case "modifier_compte":
                 if (isset($_GET["id_compte"])) {
                     if (empty($_POST["nom_compte"])) {
-                        $confirmation =
-                            "<p class='pas_ok'>Le nom du compte est obligatoire</p>";
+                        $confirmation = "<p class='pas_ok'>Le nom du compte est obligatoire</p>";
                     } elseif (empty($_POST["email_compte"])) {
                         $confirmation =
                             "<p class='pas_ok'>Le champ email est obligatoire</p>";
@@ -125,28 +110,16 @@ if (isset($_SESSION["id_compte"])) {
                     } else {
                         $requete =
                             "UPDATE comptes SET 
-                        nom_compte='" .
-                            security($_POST["nom_compte"]) .
-                            "',
-                        prenom_compte='" .
-                            security($_POST["prenom_compte"]) .
-                            "',
-                        email_compte='" .
-                            security($_POST["email_compte"]) .
-                            "',
-                        login_compte='" .
-                            security($_POST["login_compte"]) .
-                            "'";
+                        nom_compte='" . security($_POST["nom_compte"]) . "',
+                        prenom_compte='" . security($_POST["prenom_compte"]) . "',
+                        email_compte='" . security($_POST["email_compte"]) . "',
+                        login_compte='" . security($_POST["login_compte"]) . "'";
 
                         if (!empty($_POST["pass_compte"])) {
                             // modified password
-                            $requete .=
-                                ",pass_compte=SHA1('" .
-                                $_POST["pass_compte"] .
-                                "')";
+                            $requete .= ",pass_compte=SHA1('" . $_POST["pass_compte"] . "')";
                         }
-                        $requete .=
-                            " WHERE id_compte='" . $_GET["id_compte"] . "'";
+                        $requete .= " WHERE id_compte='" . $_GET["id_compte"] . "'";
                         $resultat = mysqli_query($connexion, $requete);
                     }
 
@@ -154,8 +127,7 @@ if (isset($_SESSION["id_compte"])) {
                         unset($_POST[$cle]);
                     }
 
-                    $confirmation =
-                        "<p class='ok'>Le compte a bien été modifié ! </p> ";
+                    $confirmation = "<p class='ok'>Le compte a bien été modifié ! </p> ";
                 }
 
                 break;
@@ -171,26 +143,15 @@ if (isset($_SESSION["id_compte"])) {
         $content .= "<details class=\"tab_results\">";
 
         $content .= "<summary>";
+
         $content .= "<div>" . $ligne->id_compte . "</div>";
+        $content .= "<div>" . $ligne->statut_compte . "</div>";
         $content .= "<div>" . $ligne->login_compte . "</div>";
         $content .= "<div>" . $ligne->email_compte . "</div>";
-        $content .=
-            "<div><a href=\"back.php?action=compte&cas=recharger_compte&id_compte=" .
-            $ligne->id_compte .
-            "\">modifier</a></div>";
-        $content .=
-            "<div><a href=\"back.php?action=compte&cas=avertir_compte&id_compte=" .
-            $ligne->id_compte .
-            "\">supprimer</a></div>";
+        $content .="<div><a href=\"back.php?action=compte&cas=recharger_compte&id_compte=" . $ligne->id_compte . "\">modifier</a></div>";
+        $content .= "<div><a href=\"back.php?action=compte&cas=avertir_compte&id_compte=" . $ligne->id_compte . "\">supprimer</a></div>";
         $content .= "</summary>";
-
-        $content .=
-            "<div>" .
-            $ligne->nom_compte .
-            " " .
-            $ligne->prenom_compte .
-            "</div>";
-
+        $content .="<div>" . $ligne->nom_compte . " " . $ligne->prenom_compte . "</div>";
         $content .= "</details>";
     }
 } else {
